+++ src/schema/__init__.py
@@ -1,11 +1,12 @@
 from .display_triple import DisplayTriple
 from .endpoint import Endpoint
 from .entity import Entity
-from .enums import EndpointTechnology, OntologyFramework
+from .enums import EndpointTechnology, OntologyFramework, EntityType
 from .graph import Graph
 from .ontology import Ontology
 from .ontology_class import OntologyClass
 from .ontology_property import OntologyProperty
+from .predicate import Predicate
 from .prefix import Prefix
 from .query import Query
 from .triple import Triple